place test failure warning beside invocation error
